[GLUTEN-5307][VL] Fix Potential Overflow Issue in VeloxShuffleWriter Due to Mismatched Data Types of RowNumber

+++ cpp/core/shuffle/RoundRobinPartitioner.h
@@ -30,7 +30,7 @@ class RoundRobinPartitioner final : public Partitioner {
       const int32_t* pidArr,
       const int64_t numRows,
       std::vector<uint32_t>& row2Partition,
-      std::vector<uint16_t>& partition2RowCount) override;
+      std::vector<uint32_t>& partition2RowCount) override;
 
  private:
   friend class RoundRobinPartitionerTest;
